// App/web/src/lib/generated-api/band-api/models/member.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * BandService | v1
 * OpenAPI spec version: 1.0.0
 */
import type { MemberRole } from './memberRole';
import type { Permission } from './permission';

export interface Member {
  id: string;
  userId: string;
  email: string;
  bandId: string;
  role: MemberRole;
  joinedAt: string;
  permissions?: Permission;
}